Some minor changes to the JSON format in preparation for the addition of a new field. Additionally, switched over to using f32 for the values since f64 is overkill.

+++ identify-the-file/src/main.rs
@@ -170,7 +170,7 @@ struct PatternMatch<'a> {
     pub uuid: &'a str,
     pub points: usize,
     pub max_points: usize,
-    pub percentage: f64,
+    pub percentage: f32,
 }
 
 impl<'a> PatternMatch<'a> {
@@ -179,7 +179,7 @@ impl<'a> PatternMatch<'a> {
             uuid,
             points,
             max_points,
-            percentage: utils::round_to_dp(points as f64 / max_points as f64 * 100.0, 1),
+            percentage: utils::round_to_dp(points as f32 / max_points as f32 * 100.0, 1),
         }
     }
 }
